+++ gradio/components.py
@@ -2779,6 +2779,138 @@ class Chatbot(Component):
         self.set_event_trigger("change", fn, inputs, outputs)
 
 
+class Model3D(Component):
+    """
+    Component creates a 3D Model component with input and output capabilities.
+    Input type: File object of type (.obj, glb, or .gltf)
+    Output type: filepath
+    Demos: Model3D
+    """
+
+    def __init__(
+        self, 
+        clear_color=None,         
+        label: str = None,
+        css: Optional[Dict] = None,
+        **kwargs,):
+        """
+        Parameters:
+        clear_color (List[r, g, b, a]): background color of scene
+        label (str): component name in interface.
+        """
+        self.clear_color = clear_color
+        super().__init__(label=label, css=css, **kwargs)
+
+    def get_template_context(self):
+        return {**super().get_template_context()}
+
+    @classmethod
+    def get_shortcut_implementations(cls):
+        return {
+            "Model3D": {},
+        }
+
+    def preprocess_example(self, x):
+        return {"name": x, "data": None, "is_example": True}
+
+    def preprocess(self, x: Dict[str, str] | None) -> str | None:
+        """
+        Parameters:
+        x (Dict[name: str, data: str]): JSON object with filename as 'name' property and base64 data as 'data' property
+        Returns:
+        (str): file path to 3D image model
+        """
+        if x is None:
+            return x
+        file_name, file_data, is_example = (
+            x["name"],
+            x["data"],
+            x.get("is_example", False),
+        )
+        if is_example:
+            file = processing_utils.create_tmp_copy_of_file(file_name)
+        else:
+            file = processing_utils.decode_base64_to_file(
+                file_data, file_path=file_name
+            )
+        file_name = file.name
+        return file_name
+
+    def serialize(self, x, called_directly):
+        raise NotImplementedError()
+
+    def save_flagged(self, dir, label, data, encryption_key):
+        """
+        Returns: (str) path to 3D image model file
+        """
+        return self.save_flagged_file(
+            dir, label, None if data is None else data["data"], encryption_key
+        )
+
+    def generate_sample(self):
+        return media_data.BASE64_MODEL3D
+
+    # Output functions
+
+    def postprocess(self, y):
+        """
+        Parameters:
+        y (str): path to the model
+        Returns:
+        (str): file name
+        (str): file extension
+        (str): base64 url data
+        """
+
+        if self.clear_color is None:
+            self.clear_color = [0.2, 0.2, 0.2, 1.0]
+
+        return {
+            "name": os.path.basename(y),
+            "extension": os.path.splitext(y)[1],
+            "clearColor": self.clear_color,
+            "data": processing_utils.encode_file_to_base64(y),
+        }
+
+    def deserialize(self, x):
+        return processing_utils.decode_base64_to_file(x).name
+
+    def save_flagged(self, dir, label, data, encryption_key):
+        """
+        Returns: (str) path to model file
+        """
+        return self.save_flagged_file(dir, label, data["data"], encryption_key)
+
+    def change(self, fn: Callable, inputs: List[Component], outputs: List[Component]):
+        """
+        Parameters:
+            fn: Callable function
+            inputs: List of inputs
+            outputs: List of outputs
+        Returns: None
+        """
+        self.set_event_trigger("change", fn, inputs, outputs)
+
+    def edit(self, fn: Callable, inputs: List[Component], outputs: List[Component]):
+        """
+        Parameters:
+            fn: Callable function
+            inputs: List of inputs
+            outputs: List of outputs
+        Returns: None
+        """
+        self.set_event_trigger("edit", fn, inputs, outputs)
+
+    def clear(self, fn: Callable, inputs: List[Component], outputs: List[Component]):
+        """
+        Parameters:
+            fn: Callable function
+            inputs: List of inputs
+            outputs: List of outputs
+        Returns: None
+        """
+        self.set_event_trigger("clear", fn, inputs, outputs)
+
 # Static Components
 class Markdown(Component):
     """
@@ -2970,135 +3102,3 @@ def get_component_instance(comp: str | dict | Component):
         raise ValueError(
             f"Component must provided as a `str` or `dict` or `Component` but is {comp}"
         )
-
-class Model3D(Component):
-    """
-    Component creates a 3D Model component with input and output capabilities.
-    Input type: File object of type (.obj, glb, or .gltf)
-    Output type: filepath
-    Demos: Model3D
-    """
-
-    def __init__(
-        self, 
-        clear_color=None,         
-        label: str = None,
-        css: Optional[Dict] = None,
-        **kwargs,):
-        """
-        Parameters:
-        clear_color (List[r, g, b, a]): background color of scene
-        label (str): component name in interface.
-        """
-        self.clear_color = clear_color
-        super().__init__(label=label, css=css, **kwargs)
-
-    def get_template_context(self):
-        return {**super().get_template_context()}
-
-    @classmethod
-    def get_shortcut_implementations(cls):
-        return {
-            "Model3D": {},
-        }
-
-    def preprocess_example(self, x):
-        return {"name": x, "data": None, "is_example": True}
-
-    def preprocess(self, x: Dict[str, str] | None) -> str | None:
-        """
-        Parameters:
-        x (Dict[name: str, data: str]): JSON object with filename as 'name' property and base64 data as 'data' property
-        Returns:
-        (str): file path to 3D image model
-        """
-        if x is None:
-            return x
-        file_name, file_data, is_example = (
-            x["name"],
-            x["data"],
-            x.get("is_example", False),
-        )
-        if is_example:
-            file = processing_utils.create_tmp_copy_of_file(file_name)
-        else:
-            file = processing_utils.decode_base64_to_file(
-                file_data, file_path=file_name
-            )
-        file_name = file.name
-        return file_name
-
-    def serialize(self, x, called_directly):
-        raise NotImplementedError()
-
-    def save_flagged(self, dir, label, data, encryption_key):
-        """
-        Returns: (str) path to 3D image model file
-        """
-        return self.save_flagged_file(
-            dir, label, None if data is None else data["data"], encryption_key
-        )
-
-    def generate_sample(self):
-        return media_data.BASE64_MODEL3D
-
-    # Output functions
-
-    def postprocess(self, y):
-        """
-        Parameters:
-        y (str): path to the model
-        Returns:
-        (str): file name
-        (str): file extension
-        (str): base64 url data
-        """
-
-        if self.clear_color is None:
-            self.clear_color = [0.2, 0.2, 0.2, 1.0]
-
-        return {
-            "name": os.path.basename(y),
-            "extension": os.path.splitext(y)[1],
-            "clearColor": self.clear_color,
-            "data": processing_utils.encode_file_to_base64(y),
-        }
-
-    def deserialize(self, x):
-        return processing_utils.decode_base64_to_file(x).name
-
-    def save_flagged(self, dir, label, data, encryption_key):
-        """
-        Returns: (str) path to model file
-        """
-        return self.save_flagged_file(dir, label, data["data"], encryption_key)
-
-    def change(self, fn: Callable, inputs: List[Component], outputs: List[Component]):
-        """
-        Parameters:
-            fn: Callable function
-            inputs: List of inputs
-            outputs: List of outputs
-        Returns: None
-        """
-        self.set_event_trigger("change", fn, inputs, outputs)
-
-    def edit(self, fn: Callable, inputs: List[Component], outputs: List[Component]):
-        """
-        Parameters:
-            fn: Callable function
-            inputs: List of inputs
-            outputs: List of outputs
-        Returns: None
-        """
-        self.set_event_trigger("edit", fn, inputs, outputs)
-
-    def clear(self, fn: Callable, inputs: List[Component], outputs: List[Component]):
-        """
-        Parameters:
-            fn: Callable function
-            inputs: List of inputs
-            outputs: List of outputs
-        Returns: None
-        """
-        self.set_event_trigger("clear", fn, inputs, outputs)
