+++ ui/packages/model3D/src/Model3D.svelte
@@ -1,101 +1,37 @@
 <script lang="ts">
-	import { createEventDispatcher, tick } from "svelte";
-	import { Upload, ModifyUpload } from "@gradio/upload";
 	import type { FileData } from "@gradio/upload";
-	import { Block, BlockLabel } from "@gradio/atoms";
 
-	import file_icon from "./file.svg";
-
-	export let value: null | FileData;
-
-	export let drop_text: string = "Drop a file";
-	export let or_text: string = "or";
-	export let upload_text: string = "click to upload";
-	export let label: string = "";
-	export let style: string;
-
-	let file_count: string;
-
-	async function handle_upload({ detail }: CustomEvent<FileData>) {
-		value = detail;
-		await tick();
-		dispatch("change", value);
-	}
-
-	async function handle_clear({ detail }: CustomEvent<null>) {
-		value = null;
-		await tick();
-		dispatch("clear");
-	}
-
-	const dispatch =
-		createEventDispatcher<{ change: FileData | null; clear: undefined }>();
-
-	let dragging = false;
-
-	import { afterUpdate } from "svelte";
-	import * as BABYLON from "babylonjs";
-	import "babylonjs-loaders";
-
-	let canvas: HTMLCanvasElement;
-	let scene: BABYLON.Scene;
-
-	afterUpdate(() => {
-		if (value != null) {
-			addNewModel();
-		}
-	});
-
-	function addNewModel() {
-		const engine = new BABYLON.Engine(canvas, true);
-		scene = new BABYLON.Scene(engine);
-		scene.createDefaultCameraOrLight();
-		scene.clearColor = new BABYLON.Color4(0.2, 0.2, 0.2, 1);
-		engine.runRenderLoop(() => {
-			scene.render();
-		});
-
-		window.addEventListener("resize", () => {
-			engine.resize();
-		});
-
-		let base64_model_content = value.data;
-		let raw_content = BABYLON.Tools.DecodeBase64(base64_model_content);
-		let blob = new Blob([raw_content]);
-		let url = URL.createObjectURL(blob);
-		BABYLON.SceneLoader.Append(
-			"",
-			url,
-			scene,
-			() => {
-				scene.createDefaultCamera(true, true, true);
-			},
-			undefined,
-			undefined,
-			"." + value.name.split(".")[1]
-		);
-	}
+	export let value: FileData;
+	export let style: string = "";
 </script>
 
-<Block
-	variant={value === null ? "dashed" : "solid"}
-	color={dragging ? "green" : "grey"}
-	padding={false}
+<a
+	class="output-file w-full h-full flex flex-row flex-wrap justify-center items-center relative"
+	href={value.data}
+	download={value.name}
 >
-	<BlockLabel image={file_icon} label={label || "3D Model File"} />
+	<svg
+		xmlns="http://www.w3.org/2000/svg"
+		class="h-10 w-1/5"
+		fill="none"
+		viewBox="0 0 24 24"
+		stroke="currentColor"
+	>
+		<path
+			stroke-linecap="round"
+			stroke-linejoin="round"
+			stroke-width="2"
+			d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z"
+		/>
+	</svg>
+	<div class="file-name w-3/5 text-4xl p-6 break-all">{value.name}</div>
+	<div class="text-2xl p-2">
+		{isNaN(value.size || NaN) ? "" : prettyBytes(value.size || 0)}
+	</div>
+</a>
 
-	{#if value === null}
-		{console.log("###########12312312313")}
-		<Upload on:load={handle_upload} filetype=".obj, .gltf, .glb" bind:dragging>
-			{drop_text}
-			<br />- {or_text} -<br />
-			{upload_text}
-		</Upload>
-	{:else}
-		<div
-			class="input-model w-full h-60 flex justify-center items-center bg-gray-200 dark:bg-gray-600 relative"
-		>
-			<canvas class="w-full h-full object-contain" bind:this={canvas} />
-		</div>
-	{/if}
-</Block>
+<style lang="postcss">
+	.output-file[theme="default"] {
+		@apply h-60 hover:text-gray-500;
+	}
+</style>
