+++ ui/packages/model3D/src/index.ts
@@ -1,1 +1,2 @@
-export { default as Model3D } from "./Model3D.svelte";
+export { default as Model3D } from "./Model3DUpload.svelte";
+export { default as Model3DUpload } from "./Model3DUpload.svelte";
